--- front-end/src/app/gpt-histories/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Gpthistories = () => {
-  return (
-    <div>
-      <h2>gpt 기록</h2>
-    </div>
-  );
-};
-
-export default Gpthistories;
