--- front-end/src/app/memo/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Memo = () => {
-  return (
-    <div>
-      <h2>메모</h2>
-    </div>
-  );
-};
-
-export default Memo;
