--- front-end/src/app/search-histories/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Searchhistories = () => {
-  return (
-    <div>
-      <h2>검색기록</h2>
-    </div>
-  );
-};
-
-export default Searchhistories;
